labeled-response: add ACK message
The ACK message may be used in place of an empty batch when
no response is required. This is desirable if clients blindly
label all outgoing messages, as empty batches are a bit
verbose for common and high volume operations such as PONG
responses or PRIVMSG messages when echo-message isn't negotiated.